Eliminated some threading, fixed writing to actuator board
I got rid of the packetqueue business, because although it's somewhat
less safe now, it's just so much less of a headache than worrying
about threading.

Also, the system for writing to the actuator boards was messed up -- I
had things being written to the FT245 board, which just sets the port,
instead of the F232 board, which is the way it's supposed to
happen. (I don't understand that part of the system very well, it
seems.)