Reorganized actuators and tracker to use py4j; minor bug fixes.
I found py4j, a Python package that allows you to interface with a
Java virtual machine. It seems to install easily, so I used it to fix
the two biggest warts on the program: the
calling-Java-FARO-stuff-from-Python setup and the
the-Python-D2XX-library-sucks-intensely problem. Both seem to work, at
least as far as I can test on my home computer.