Added the MeasureFrame class, started using NamedEntries, misc. minor changes.
The MeasureFrame is another tracker-related frame available in the main window. It gives the user finer control over how data are taken from the tracker, and lets them dump it into a file.

NamedEntryFrames are provided by srptools.tkinter, and are just frames
with a bunch of labeled fields for entering text within them.