Changed everything to use Vectors; added plotting; changed Tracker datapoints.
As much as possible now, everything tracker-related uses Vectors now,
rather than (r, theta, phi) tuples. They're just handier to manipulate.

There's a PlottingFrame GUI component now, which is basically just a
MeasureFrame with a "plot" button on the bottom which plots the data
instead of dumping it into a file.

I also added PyMMT.tracker.DataPoint, which we use instead of the raw
Java MeasureDataPoint class, largely because it can convert the position information to a Vector.